third_party/blink/web_tests/fast/xpath/4XPath/Borrowed/kd_20010423-expected.txt

PASS actual is "abcabcabc"
PASS successfullyParsed is true

TEST COMPLETE
